requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
mysqlclient==2.0.1
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.4.1
